// integration-testing/types/execute_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type ExecuteMsg =
  | {
      join_game: {
        num_hands_to_win: number;
        [k: string]: unknown;
      };
    }
  | {
      leave_waiting_queue: {
        [k: string]: unknown;
      };
    }
  | {
      commit_move: {
        hashed_move: string;
        player1: string;
        player2: string;
        [k: string]: unknown;
      };
    }
  | {
      reveal_move: {
        game_move: GameMove;
        nonce: string;
        player1: string;
        player2: string;
        [k: string]: unknown;
      };
    }
  | {
      claim_game: {
        player1: string;
        player2: string;
        [k: string]: unknown;
      };
    }
  | {
      forfeit_game: {
        [k: string]: unknown;
      };
    }
  | {
      update_admin: {
        admin?: string | null;
        [k: string]: unknown;
      };
    };
export type GameMove = "Rock" | "Paper" | "Scissors";