// integration-testing/types/query_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type QueryMsg =
  | {
      get_game_by_player: {
        player: string;
        [k: string]: unknown;
      };
    }
  | {
      get_game_by_players: {
        player1: string;
        player2: string;
        [k: string]: unknown;
      };
    }
  | {
      get_leaderboard: {
        limit?: number | null;
        start_after?: string | null;
        [k: string]: unknown;
      };
    }
  | {
      get_open_games: {
        limit?: number | null;
        start_after?: string | null;
        [k: string]: unknown;
      };
    }
  | {
      get_games: {
        limit?: number | null;
        start_after?: string | null;
        [k: string]: unknown;
      };
    }
  | {
      admin: {
        [k: string]: unknown;
      };
    };